### 2023-08-01 #5722 @editorialbot
[PRE REVIEW]: helios: An R package to process heating and cooling degrees for GCAM
**Submitting author:** @mengqi-z (<a href="http://orcid.org/0000-0001-5385-2758">Mengqi Zhao</a>)
**Repository:** https://github.com/JGCRI/helios
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
