### 2023-08-01 #5722 @kthyng
Note to future editor: you might consider having a non-conflicted author of this submission review: https://github.com/openjournals/joss-reviews/issues/5724
