### 2023-11-07 #5722 @ibarraespinosa
Hi, I can help review this paper/software, but I will have time only after AGU.

Best
