### 2023-11-10 #6033 @RBhupi
@mengqi-z I am trying a quick installation and having following dependency issue. Please, check.

```
> devtools::install_github("JGCRI/helios")
Downloading GitHub repo JGCRI/helios@HEAD
Downloading GitHub repo JGCRI/rmap@HEAD
Skipping 1 packages not available: jgcricolors
── R CMD build ────────────────────────────────────────────────────────────
✔  checking for file ‘/private/var/folders/8t/ld6_fkgd62q7qm1kqd0q_h5r0000gp/T/RtmpUItgiR/remotes1634f5abf84d5/JGCRI-rmap-2e357fa/DESCRIPTION’ ...
─  preparing ‘rmap’:
     NB: this package now depends on R (>= 3.5.0)
     WARNING: Added dependency on R >= 3.5.0 because serialized objects in
     serialize/load version 3 cannot be read in older versions of R.
     File(s) containing such objects:
       ‘rmap/data/allSubRegions.rda’ ‘rmap/data/exampleMapDataClass.rda’
       ‘rmap/data/exampleMapDataParam.rda’
       ‘rmap/data/example_gridData_GWPv4To2015.rda’
       ‘rmap/data/mapCountries.rda’ ‘rmap/data/mapCountriesUS52.rda’
       ‘rmap/data/mapGCAMBasins.rda’ ‘rmap/data/mapGCAMBasinsUS49.rda’
       ‘rmap/data/mapGCAMBasinsUS52.rda’ ‘rmap/data/mapGCAMLand.rda’
       ‘rmap/data/mapGCAMReg32.rda’ ‘rmap/data/mapGCAMReg32EU.rda’
       ‘rmap/data/mapGCAMReg32US52.rda’ ‘rmap/data/mapGCAMReg32Uruguay.rda’
       ‘rmap/data/mapHydroShed1.rda’ ‘rmap/data/mapHydroShed2.rda’
       ‘rmap/data/mapHydroShed3.rda’
       ‘rmap/data/mapIntersectGCAMBasin32RegUruguay.rda’
       ‘rmap/data/mapIntersectGCAMBasinCountry.rda’
       ‘rmap/data/mapIntersectGCAMBasinUS52.rda’
       ‘rmap/data/mapIntersectGCAMBasinUS52County.rda’
       ‘rmap/data/mapStates.rda’ ‘rmap/data/mapUS49.rda’
       ‘rmap/data/mapUS49County.rda’ ‘rmap/data/mapUS49HUC2.rda’
       ‘rmap/data/mapUS49HUC4.rda’ ‘rmap/data/mapUS52.rda’
       ‘rmap/data/mapUS52Compact.rda’ ‘rmap/data/mapUS52County.rda’
       ‘rmap/data/mapUS52CountyCompact.rda’ ‘rmap/data/mapUS52HUC2.rda’
       ‘rmap/data/mapUS52HUC4.rda’ ‘rmap/data/mapping_US49.rda’
       ‘rmap/data/mapping_US52.rda’ ‘rmap/data/mapping_country_gcam32.rda’
       ‘rmap/data/mapping_gcambasins.rda’
       ‘rmap/data/mapping_tethys_grid_basin_region_country.rda’
       ‘rmap/inst/extras/rmap_tethys_grid_basin_region_country.rds’
─  building ‘rmap_1.0.0.tar.gz’

ERROR: dependency ‘jgcricolors’ is not available for package ‘rmap’
* removing ‘/Library/Frameworks/R.framework/Versions/4.3-x86_64/Resources/library/rmap’
Downloading GitHub repo JGCRI/rmap@HEAD
Skipping rmap, it is already being installed
── R CMD build ────────────────────────────────────────────────────────────
✔  checking for file ‘/private/var/folders/8t/ld6_fkgd62q7qm1kqd0q_h5r0000gp/T/RtmpUItgiR/remotes1634f26d99c71/JGCRI-helios-2bc630c/DESCRIPTION’ ...
─  preparing ‘helios’:
     NB: this package now depends on R (>= 3.5.0)
     WARNING: Added dependency on R >= 3.5.0 because serialized objects in
     serialize/load version 3 cannot be read in older versions of R.
     File(s) containing such objects:
       ‘helios/data/L244.HDDCDD_building.rda’
       ‘helios/data/L2441.HDDCDD_Fixed_gcamusa_seg.rda’
       ‘helios/data/example_cmip6_china_ncdf.rda’
       ‘helios/data/example_hdcd_monthly_usa.rda’
       ‘helios/data/example_hdcd_segment_usa.rda’
       ‘helios/data/example_pop_china_ncdf.rda’
       ‘helios/data/example_pop_usa_csv.rda’
       ‘helios/data/example_wrf_usa_ncdf.rda’ ‘helios/data/mapUS52.rda’
       ‘helios/data/noaa_hddcdd.rda’ ‘helios/data/segment_map_utc.rda’
       ‘helios/data/spatial_options.rda’
       ‘helios/inst/extras/noaa_hddcdd.RDS’
─  building ‘helios_1.0.0.tar.gz’

ERROR: dependency ‘rmap’ is not available for package ‘helios’
* removing ‘/Library/Frameworks/R.framework/Versions/4.3-x86_64/Resources/library/helios’
Warning messages:
1: In i.p(...) :
  installation of package ‘/var/folders/8t/ld6_fkgd62q7qm1kqd0q_h5r0000gp/T//RtmpUItgiR/file1634f53a9ea7e/rmap_1.0.0.tar.gz’ had non-zero exit status
2: In i.p(...) :
  installation of package ‘/var/folders/8t/ld6_fkgd62q7qm1kqd0q_h5r0000gp/T//RtmpUItgiR/file1634f77f716b/helios_1.0.0.tar.gz’ had non-zero exit status
>
```
