### 2023-11-13 #6033 @mengqi-z
Hi @RBhupi , thank you for your feedback. This error is caused from one of `helios` dependencies called `rmap`. `rmap` has a dependency called `jgcricolors` that was removed from CRAN, which causes the installation issue. I updated the `rmap` so that it will install `jgcricolors` directly from its github. This should solve the problem you had. Could you please try again? Thank you!
