### 2023-12-14 #6033 @hugoledoux
No! It just means that if there are claims of novelty for a method/algo you can as a reader reproduce them, but if not then just click it. Same for data, if all data necessary for you to test the package are available then you can click it.
