### 2023-12-31 JGCRI/helios#9 @ibarraespinosa
review for JOSS
Here is my review for JOSS https://github.com/openjournals/joss-reviews/issues/6033

fyi, i'm using this system
![Screenshot_20231231-121149.png](https://github.com/JGCRI/helios/assets/27447280/808ad448-18ed-4834-97f6-2359aa51350d)
