### 2023-12-31 JGCRI/helios#10 @ibarraespinosa
review for joss
Hi

I'm reviewing for JOSS here https://github.com/openjournals/joss-reviews/issues/6033

Im using the configuration show in the picture (android), but may use other tools ![Screenshot_20231231-121149.png](https://github.com/JGCRI/helios/assets/27447280/1a311a19-c3a6-4853-8f0e-377e993aaef0)
