### 2024-01-01 JGCRI/helios#10 @ibarraespinosa
Overview:

I'm running all examples sucessfully.
This package reads surface temperature from atmospheric transport models in NetCDF format.
The author was able to create a [function](https://jgcri.github.io/helios/reference/read_ncdf.html) to read NetCDF relaying on raster.
The author preferred [wide](https://tidyr.tidyverse.org/reference/pivot_longer.html) format to process data.frames, dor instance:

```
# example data: WRF hourly climate at 12-km resolution
path_to_climate_ncdf <- helios::pkg_example('wrfout_d01_2020-01-01_01%3A00%3A00_sub.nc')

temperature <- helios::read_ncdf(ncdf = path_to_climate_ncdf,
                                 model = 'wrf',
                                 var = 'T2',
                                 time_periods = 2020)
```
The objective is to calculate temperatures outside comfort. The (figure)[https://jgcri.github.io/helios/articles/vignetteFigs/helios_gcam-usa_workflow.jpg] should be in Celsius or Fahrenheit, but calculations in kelvin.
The author may include units::set_units() to control conversions of temperature units
The code in `helios::hdcd` looks sound. I'm not a fan of dplyr, I prefer data.table, but code looks ok.
The documentation is (incomplete)[https://jgcri.github.io/helios/reference/hdcd.html]. For instance, the argument `reference_temp_F` is 'Default = 65' which is not explanatory. Does that mean that the temperature inputted must be in Fahrenheit?
Global Change Assessment Model (GCAM) should be defined in the README (first page)
In general, everything looks fine
