### 2024-01-01 JGCRI/helios#10 @ibarraespinosa
A link to (CONTRIBUTING)[https://jgcri.github.io/helios/CONTRIBUTING.html] should be added in README
