### 2024-01-01 JGCRI/helios#10 @ibarraespinosa
Compile the paper https://github.com/JGCRI/helios/blob/main/paper/paper.md (use rticles) and upload the pdf in the repo
