### 2024-01-01 JGCRI/helios#10 @ibarraespinosa
The caption of Figure 1 is too limited. Should expand it
