### 2024-01-01 JGCRI/helios#10 @ibarraespinosa
Can you change units of figure 1 from kelvin to Celsius or F?
