### 2024-01-01 JGCRI/helios#10 @ibarraespinosa
Can you add this section into the manuscript?
[] State of the field: Do the authors describe how this software compares to other commonly-used packages?
