### 2024-01-06 2b881bc
Update paper to include state of  the field
