+++ paper/paper.md
@@ -33,11 +33,15 @@ bibliography: paper.bib
 
 `helios` was developed to meet the increasing research interests to explore the spatial and temporal heterogeneity of climate impacts on sub-annual electricity demand from buildings. @ciscar2014integrated pointed out most integrated modeling systems designed to link human-Earth systems are unable to take advantage of the publicly available high resolution data to account for the impact of seasonal temperature change on energy system. To better fill in this gap, researchers have developed GCAM versions (e.g., GCAM-USA) to include power sector details at sub-annual and sub-national level [@wise2019representing]. For example, @khan2021impacts used GCAM-USA to show that the temperature-induced heating and cooling demands can significantly affect sub-annual electricity demand profiles and peak electricity loads. Understanding the seasonal dynamics of electricity demand and capacity within IAMs is of importance to support future infrastructure planning [@binsted2022electrified]. We develop `helios` to bridge the gap between high resolution data and global scale models by facilitating the workflow in estimating population-weighted heating and cooling degrees. `helios` serves as a pre-processing tool of GCAM for researchers to capture the impact of sub-annual variation of different climate and socioeconomic scenarios on building energy demand.
 
+# Statement of Field
+
+Heating and cooling degree days are commonly used as meteorological indices in the energy system to measure the temperature deviations from the reference temperature over time. These indices are widely calculated at point scale rather than spatial scale. With the increasing availability of spatially distributed climate data, few tools are developed to access and post-process it from the original climate data format into tabular format, such as `Climate4R` [@iturbide2019].However, there are rarely well-documented and open-source tools that streamline the calculation of population-weighted HDD and CDD at user-defined spatiotemporal resolutions or electricity dispatch sectors defined in GCAM, directly using gridded climate and population data. `helios` is developed to integrate these workflows and standardize the output for easy usage within and beyond GCAM applications.
+
 # Design and Functionality
 
 `helios` is designed to provide heating and cooling degrees to GCAM (or GCAM-USA) at two spatiotemporal scales: (1) HDH and CDH for the U.S. States for dispatch segments by building thermal service (for GCAM-USA); (2) annual HDD and CDD for GCAM's 32 geopolitical regions at 5-year time step by building thermal service (for GCAM-Regions). Beyond providing information for GCAM, `helios` can serve as a general tool to calculate heating and cooling degrees at a monthly time step and various spatial scales (e.g., country, basin). The use of `helios` requires users to provide information about the input files, such as the climate model source for climate data and the variable name for temperature. For example, the ISIMIP-CMIP data uses "tas" for the variable name for temperature while the WRF data uses "T2". Figure 1 shows the workflow for both GCAM-USA and GCAM-Regions. More details can be accessed in the `helios` [documentation](https://jgcri.github.io/helios/index.html) page on Github.
 
-![The workflow for helios.\label{fig:1}](Fig1_helios_workflow.jpg)
+![The Helios workflow using two input data examples of global and CONUS data. Helios takes in various temperature and population data at regional or global scales and processes the data into population-weighted heating and cooling degrees at a user-defined spatiotemporal scale. \label{fig:1}](Fig1_helios_workflow.jpg)
 
 Working with climate data can pose challenges, given large data sizes and diverse formats, spatiotemporal resolutions, data structures, and dimensions involved. The `helios` package provides functionality that makes it more convenient for users to manipulate climate data. `helios` provides easier access to various climate data types in a simplified format, facilitates the calculation of heating and cooling degrees using a standardized methodology, and ensures quality control through detailed diagnostics. There are five main functions provided by `helios`:
 
