### 2024-01-09 ecf5239
label units for temperature and HDD/CDD/HDH/CDH
1. for read_ncdf, add unit of the dataset in metadata
2. for hdcd output, modify the unit column to indicate Fahrenheit degree-days/hours
