+++ paper/paper.md
@@ -35,13 +35,13 @@ bibliography: paper.bib
 
 # Statement of Field
 
-Heating and cooling degree days are commonly used as meteorological indices in the energy system to measure the temperature deviations from the reference temperature over time. These indices are widely calculated at point scale rather than spatial scale. With the increasing availability of spatially distributed climate data, few tools are developed to access and post-process it from the original climate data format into tabular format, such as `Climate4R` [@iturbide2019].However, there are rarely well-documented and open-source tools that streamline the calculation of population-weighted HDD and CDD at user-defined spatiotemporal resolutions or electricity dispatch sectors defined in GCAM, directly using gridded climate and population data. `helios` is developed to integrate these workflows and standardize the output for easy usage within and beyond GCAM applications.
+Heating and cooling degree days are commonly used as meteorological indices in the energy system to measure the temperature deviations from the reference temperature over time. These indices are widely calculated at point scale rather than spatial scale. With the increasing availability of spatially distributed climate data, few tools are developed to access and post-process the raw climate data format into tabular format, such as `Climate4R` [@iturbide2019].However, there are rarely well-documented and open-source tools that streamline the calculation of population-weighted HDD and CDD at user-defined spatiotemporal resolutions or electricity dispatch sectors defined in GCAM, directly using gridded climate and population data. `helios` is developed to integrate these workflows and standardize the output for easy usage within and beyond GCAM applications.
 
 # Design and Functionality
 
 `helios` is designed to provide heating and cooling degrees to GCAM (or GCAM-USA) at two spatiotemporal scales: (1) HDH and CDH for the U.S. States for dispatch segments by building thermal service (for GCAM-USA); (2) annual HDD and CDD for GCAM's 32 geopolitical regions at 5-year time step by building thermal service (for GCAM-Regions). Beyond providing information for GCAM, `helios` can serve as a general tool to calculate heating and cooling degrees at a monthly time step and various spatial scales (e.g., country, basin). The use of `helios` requires users to provide information about the input files, such as the climate model source for climate data and the variable name for temperature. For example, the ISIMIP-CMIP data uses "tas" for the variable name for temperature while the WRF data uses "T2". Figure 1 shows the workflow for both GCAM-USA and GCAM-Regions. More details can be accessed in the `helios` [documentation](https://jgcri.github.io/helios/index.html) page on Github.
 
-![The Helios workflow using two input data examples of global and CONUS data. Helios takes in various temperature and population data at regional or global scales and processes the data into population-weighted heating and cooling degrees at a user-defined spatiotemporal scale. \label{fig:1}](Fig1_helios_workflow.jpg)
+![An example of the Helios workflow using two types of input datasets (e.g., global data from CMIP6 and CONUS data from WRF). This demonstration showcases Helios' capability to generate heating and cooling degrees by GCAM region or U.S. States, among other spatiotemporal scales. \label{fig:1}](Fig1_helios_workflow.jpg)
 
 Working with climate data can pose challenges, given large data sizes and diverse formats, spatiotemporal resolutions, data structures, and dimensions involved. The `helios` package provides functionality that makes it more convenient for users to manipulate climate data. `helios` provides easier access to various climate data types in a simplified format, facilitates the calculation of heating and cooling degrees using a standardized methodology, and ensures quality control through detailed diagnostics. There are five main functions provided by `helios`:
 
