### 2024-01-09 JGCRI/helios#10 @mengqi-z
Hi @ibarraespinosa,

Thank you for your valuable comments and suggestions. We have addressed them, and you can review all the commits summarized in the pull request https://github.com/JGCRI/helios/pull/12.

Please note that we didn't change the temperature unit from "K" to "Celsius" or "Fahrenheit" in the workflow figure because Kelvin is the unit of the raw temperature data for both CMIP and WRF cases. After reading the raw temperature data, `helios` will convert the temperature from Kelvin to Fahrenheit before calculating heating and cooling degrees.

We appreciate the reviewer's suggestion of using `units::set_units()` to clarify the output data units. However, we believed that having the units directly associated with the values could make later calculations a bit challenging for users who wish to manipulate the outputs. Instead, we made the following updates:

- We added an attribute to the output table from `helios::read_ncdf` to indicate the unit of the raw temperature data. Users can check the unit information using `str(output)` or `attr(output, 'metadata')`.

- We also updated the unit column in the output table from `helios::hdcd` to indicate that the degree-days/hours are Fahrenheit-based.

Thank you once again for your review of `helios`. Please don't hesitate to let us know if you have any further comments or questions.
