+++ pkgdown.yml
@@ -4,5 +4,5 @@ pkgdown_sha: ~
 articles:
   vignette_gcam-regions: vignette_gcam-regions.html
   vignette_gcam-usa: vignette_gcam-usa.html
-last_built: 2024-01-09T16:03Z
+last_built: 2024-01-09T16:40Z
 
