### 2024-01-11 JGCRI/helios#10 @ibarraespinosa
Thanks for the change in the code, but the documentation has to improve as well.  [here](https://jgcri.github.io/helios/reference/hdcd.html), please include a better documentation for the argument. reference_temp_F
