### 2024-01-11 1a68d6b
Update function argument documentation
