### 2024-01-11 JGCRI/helios#10 @mengqi-z
Hi @ibarraespinosa, Thank you for catching that. I forgot to document the package before pushing to Github. Now all the arguments documentation should be updated accordingly.
