### 2024-02-12 #6033 @hugoledoux
@mengqi-z I have proof-read the paper and made a PR with two small fixes: https://github.com/JGCRI/helios/pull/17 Also, at l.34, the acronym "IAM" has never been introduced. Can you please fix those?

After this point could you:
