+++ R/hdcd.R
@@ -16,6 +16,7 @@
 #' @param xml Default = FALSE. Set to TRUE to generate XML outputs for GCAM.
 #' @param save Default = TRUE. Set to TRUE to save outputs.
 #' @param name_append Default = ''. String for the name to append to output file name.
+#' @param im3_analysis Default = T. Output annual HDCD at grid region scale for trend-representative year analysis
 #' @importFrom magrittr %>%
 #' @importFrom data.table :=
 #' @export
@@ -33,7 +34,8 @@ hdcd <- function(ncdf = NULL,
                  diagnostics = F,
                  xml = F,
                  name_append = '',
-                 save = T) {
+                 save = T,
+                 im3_analysis = T) {
 
   print('Starting function process_hdcd...')
 
@@ -65,7 +67,15 @@ hdcd <- function(ncdf = NULL,
     hdcd_comb_annual <- tibble::tibble()
     hdcd_region_bld <- tibble::tibble()
     hdcd_region_monthly <- tibble::tibble()
-    hdcd_region_annual<- tibble::tibble()
+    hdcd_region_annual <- tibble::tibble()
+
+    # for IM3 grid region analysis
+    if(im3_analysis){
+      hdcd_comb_gridregion_monthly <- tibble::tibble()
+      hdcd_comb_gridregion_annual <- tibble::tibble()
+    }
+
+
 
     # check population input format
     if (any(grepl('tbl_df|tbl|data.frame', class(population)))) {
@@ -208,6 +218,19 @@ hdcd <- function(ncdf = NULL,
               dplyr::ungroup() %>%
               dplyr::mutate(pop_weight = value / subRegion_total_value)
 
+            # for IM3 grid region analysis
+            if(im3_analysis){
+              population_j_weighted_gridregion <- population_j_grid %>%
+                dplyr::left_join(helios::mapping_states_gridregion, by = 'subRegion') %>%
+                dplyr::mutate(subRegion = grid_region,
+                              ID = grid_region) %>%
+                dplyr::select(-grid_region) %>%
+                dplyr::group_by(region, ID, subRegion, year) %>%
+                dplyr::mutate(subRegion_total_value = sum(value, na.rm = T)) %>%
+                dplyr::ungroup() %>%
+                dplyr::mutate(pop_weight = value / subRegion_total_value)
+            }
+
             print('Completed population weighting.')
 
             #......................
@@ -222,6 +245,23 @@ hdcd <- function(ncdf = NULL,
                                by = c('ID', 'region', 'subRegion', 'lat', 'lon', 'year')) %>%
               dplyr::mutate(value = (((value - 273.15) * 9/5) + 32) - reference_temp_F,
                             value = dplyr::if_else(is.na(pop_weight), value, value * pop_weight))
+
+            # for IM3 grid region analysis
+            if(im3_analysis){
+
+              ncdf_pivot_gridregion <- ncdf_pivot %>%
+                dplyr::left_join(helios::mapping_states_gridregion, by = 'subRegion') %>%
+                dplyr::mutate(subRegion = grid_region,
+                              ID = grid_region) %>%
+                dplyr::select(-grid_region)
+
+              ncdf_hdcd_pop_weighted_gridregion <- ncdf_pivot_gridregion %>%
+                dplyr::left_join(population_j_weighted_gridregion %>%
+                                   dplyr::select(-value, -subRegion_total_value),
+                                 by = c('ID', 'region', 'subRegion', 'lat', 'lon', 'year')) %>%
+                dplyr::mutate(value = (((value - 273.15) * 9/5) + 32) - reference_temp_F,
[…]
