+++ man/mapping_states_gridregion.Rd
@@ -1,0 +1,25 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/data.R
+\docType{data}
+\name{mapping_states_gridregion}
+\alias{mapping_states_gridregion}
+\title{mapping_states_gridregion}
+\format{
+tibble::tribble
+}
+\source{
+gcamdata/inst/extdata/gcam-usa/states_subregions.csv
+}
+\usage{
+mapping_states_gridregion
+}
+\description{
+mapping_states_gridregion
+}
+\examples{
+\dontrun{
+ library(helios);
+ mapping_states_gridregion <- helios::mapping_states_gridregion
+}
+}
+\keyword{datasets}
