### 2024-02-19 1b3b0a6
Update diagnostics for monthly outputs
