### 2024-02-22 #6033 @hugoledoux
@mengqi-z thanks, sorry for small delay in processing this. One of our EiC will now take over and might ask for some more (minor) changes.

Thanks @ibarraespinosa and @RBhupi for the reviews! I hope you enjoyed it and are up for new JOSS reviews in the future.
